+++ src/lib.rs
@@ -1,7 +1,6 @@
 pub use config::Config;
 pub use server::Server;
 
-mod event;
 mod kafka;
 
 pub mod config;