+++ CHANGELOG.md
@@ -6,12 +6,13 @@
 
 * Fix hanging requests when sending newline-delimited json with empty request body or delayed
   next chunk when using chunked-transfer encoding, by fixing the request termination/shutdown logic
+* Fix potentially uncaught errors on tokio tasks for sending messages to kafka
 
 ## 0.12.0 - 2025-01-14
 
 ### Changed
 
-* Changed configuration to support arbitrary librdkafka properties read from a file, not just 
+* Change configuration to support arbitrary librdkafka properties read from a file, not just
   sasl password
 
 ## 0.11.0 - 2025-01-13